kernel: diagnose 10 does not release memory above 2GB
Patchwork-id: 35995
O-Subject: [RHEL6.2 PATCH 06/14] [s390x] kernel: diagnose 10 does not release memory above 2GB
Bugzilla: 701274

Description
-----------
The working set size of active pages as reported by z/VM is
not consistent with the size of the CMM balloon.

The diag10 function does nothing if the address of the page
to release is higher than 2047MB because the diagnose X'10'
interface of z/VM prior to release 5.2 had been restricted
to the 31bit addressing mode and to addresses below 2047MB.

Remove addressing mode switch to 31 bit and allow addresses
above 2047MB. Add two expection table entries and continue
after the diagnose instruction in the case it causes program
checks.

Bugzilla
--------
BZ 701274
https://bugzilla.redhat.com/show_bug.cgi?id=701274

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.39
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=83ace2701b81be549cca7af33c5b0499cb2602d6

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>