dlm: make plock operation killable
Patchwork-id: 36040
O-Subject: [RHEL6.2 PATCH] dlm: make plock operation killable
Bugzilla: 707005
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

bz 707005
upstream: sent for 2.6.40
https://brewweb.devel.redhat.com/taskinfo?taskID=3343576

Allow processes blocked on plock requests to be interrupted
when they are killed.  This leaves the problem of cleaning
up the lock state in userspace.  This has three parts:

1. Add a flag to unlock operations sent to userspace
indicating the file is being closed.  Userspace will
then look for and clear any waiting plock operations that
were abandoned by an interrupted process.

2. Queue an unlock-close operation (like in 1) to clean up
userspace from an interrupted plock request.  This is needed
because the vfs will not send a cleanup-unlock if it sees no
locks on the file, which it won't if the interrupted operation
was the only one.

3. Do not use replies from userspace for unlock-close operations
because they are unnecessary (they are just cleaning up for the
process which did not make an unlock call).  This also simplifies
the new unlock-close generated from point 2.

Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>