ext4: Add mount options in superblock
Patchwork-id: 36054
O-Subject: [PATCH RHEL6.1] ext4: Add mount options in superblock
Bugzilla: 704557
RH-Acked-by: Josef Bacik <josef@redhat.com>
RH-Acked-by: Phillip Lougher <plougher@redhat.com>

This is for Bug 704557 -
[RFE] tune2fs cannot currently set barrier=0 as a default option for ext* file systems (kernel side)

It's a straightforward backport of this upstream commit.
The commit differs from upstream primarily because many intervening superblock
fields were added prior to this change, so they are included (but unused)
here.

The patch allows simple setting of default mount opts via tune2fs -o
for nobarrier, block_validity, discard, and nodelalloc.  It also allows
for arbitrary mount option strings to be stored in the sb...

Requires e2fsprogs changes too, which are committed to cvs already.

Thanks,
-Eric

commit 8b67f04ab9de5d8f3a71aef72bf02c995a506db5
Author: Theodore Ts'o <tytso@mit.edu>
Date:   Sun Aug 1 23:14:20 2010 -0400

    ext4: Add mount options in superblock

    Allow mount options to be stored in the superblock.  Also add default
    mount option bits for nobarrier, block_validity, discard, and nodelalloc.

    Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>