Fix memory leak/corruption on VLAN GRO_DROP
Patchwork-id: 4581
O-Subject: [kernel team] [RHEL6 PATCH] net: Fix memory leak/corruption on VLAN
	GRO_DROP
Bugzilla: 695176
CVE: CVE-2011-1576
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

This patch fixes a serious memory leak/corruption bug in RHEL6.
This bug used to exist upstream as well, but was fixed when the
VLAN code path was restructured in 2010.

The approach here is different from upstream in order to minimise
the effect on the code.

net: Fix memory leak/corruption on VLAN GRO_DROP

The function napi_reuse_skb is only meant to be used for packets
merged by GRO.  Using it on the VLAN path will lead to memory
leaks/corruption.  This patch is based on Jay Vosburgh's patch,
and it fixes the problem by calling kfree_skb on the VLAN GRO_DROP
path instead of napi_reuse_skb.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>