ixgbe: adding FdirMode module option
Patchwork-id: 36192
O-Subject: [PATCH RHEL6.2] ixgbe: adding FdirMode module option
Bugzilla: 707287
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

One of our customers discovered that our RHEL6 ixgbe driver did not
perform as well with their setup as it did with Intel's Sourceforge
driver.  Based on the ethtool stats (rx_missed errors and fdir_miss
stats incrementing) it did not appear the Flow Director in 82599 devices
was working as well as expected with our code.  Wondering if the latest
upstream driver worked fine and we were simply missing a few patches, I
tried it.  It was just as bad as our driver.

After talking with one of the ixgbe developers at Intel he confirmed
that there were changes in the Sourceforge driver that were still not
upstream.  While Intel figures this out we need something to support the
customer.  We discovered that disabling the Flow Director does produce
suitable performance, so adding a module option to control this seems
like the only reasonable supportable option until the real fixes hit
upstream.  I would rather not disable the Flow Director completely as it
has real benefits for most users.  This customer's unique setup is what
causes problems.

I have verified that the patch works properly on RHEL6.0.z and the
customer has also indicated it resolves their packet drop issues.

This will resolve RHBZ 707287.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>