Revert "[block] blk-throttle: Correct the placement of smp_rmb()"
Patchwork-id: 36198
O-Subject: [RHEL6.2 PATCH 01/30] Revert "[block] blk-throttle: Correct the
	placement of smp_rmb()"
Bugzilla: 707762
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

BZ:707762

This reverts commit bc100fb4307580dff6590f45bfd0731cf71f7776.

The patch which was under discussion on rhkernel-list got committed and not the
one which was upstream. First patch which was upstream was an improvement
over the state of affairs but still had issues. Later I went ahead and
put a fix in upstream to fix the issues.  So this patch never went upstream.

Hence this needs to be reverted, and first we need to commit the first
version of patch and then on top of we need to commit the additional
fix backported from upstream.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>