blk-cgroup: Make cgroup stat reset path blkg->lock free for dispatch stats
Patchwork-id: 36204
O-Subject: [RHEL6.2 PATCH 27/30] blk-cgroup: Make cgroup stat reset path
	blkg->lock free for dispatch stats
Bugzilla: 707755
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

BZ:707755

Upstream commit f0bdc8cdd9a2bcc2c84ae2a1fdbff4188b354d8d

Now dispatch stats update is lock free. But reset of these stats still
takes blkg->stats_lock and is dependent on that. As stats are per cpu,
we should be able to just reset the stats on each cpu without any locks.
(Atleast for 64bit arch).

On 32bit arch there is a small race where 64bit updates are not atomic.
The result of this race can be that in the presence of other writers,
one might not get 0 value after reset of a stat and might see something
intermediate

One can write more complicated code to cover this race like sending IPI
to other cpus to reset stats and for offline cpus, reset these directly.

Right not I am not taking that path because reset_update is more of a
debug feature and it can happen only on 32bit arch and possibility of
it happening is small. Will fix it if it becomes a real problem. For
the time being going for code simplicity.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>