ext4: Fix possible lost inode write in no journal mode
Patchwork-id: 36280
O-Subject: [RHEL6.2 PATCH] ext4: Fix possible lost inode write in no journal mode
Bugzilla: 699151
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Josef Bacik <josef@redhat.com>

BZ 699151
https://bugzilla.redhat.com/show_bug.cgi?id=699151

BREW 3350020
https://brewweb.devel.redhat.com/taskinfo?taskID=3350020

In the no-journal case, ext4_write_inode() will fetch the bh and call
sync_dirty_buffer() on it.  However, if the bh has already been
written and the bh reclaimed for some other purpose, AND if the inode
is the only one in the inode table block in use, then
ext4_get_inode_loc() will not read the inode table block from disk,
but as an optimization, fill the block with zero's assuming that its
caller will copy in the on-disk version of the inode.  This is not
done by ext4_write_inode(), so the contents of the inode can simply
get lost.  The fix is to use __ext4_get_inode_loc() with in_mem set to
0, instead of ext4_get_inode_loc().  Long term the API needs to be
fixed so it's obvious why latter is not safe.

Addresses-Google-Bug: #2526446

Signed-off-by: Curt Wohlgemuth <curtw@google.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>