nfs: coalesce small, page-aligned direct I/Os
Patchwork-id: 36344
O-Subject: [RHEL6.2 PATCH] BZ#694309: nfs: coalesce small,
	page-aligned direct I/Os (try #2)
Bugzilla: 694309
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

From: Badari Pulavarty <pbadari@us.ibm.com>

This is essentially the same patch as proposed earlier with some
small cleanups to the code that David suggested. Original description
follows.

The NFS client's DIO code is somewhat peculiar in that it handles
multiple iovecs as individual I/O's even when they could be coalesced.
This is rather inefficient when userspace sends down multiple, small
iovecs in a single readv or writev call. The kernel will issue a
separate I/O for each one.

Mostly, this is due to the fact that the DIO code in NFS was originally
written to satisfy Oracle DB usage patterns which apparently don't
involve readv/writev calls with small iovecs.

Older versions of QEMU allocated bounce buffers to send I/O's down to
the kernel. While inefficient, that had the side benefit of coalescing
the page-sized and aligned buffers that it was using. When QEMU was
changed not to use bounce buffers with O_DIRECT on NFS, the performance
fell off a cliff.

This patch is intended to fix this problem by adding a set of special
cases to the DIO codepaths in NFS. It first checks to see if the iovec
array is a candidate for the new codepath. If it fails any of the checks
then it will fall back to using the old code. Ditto if any allocations
fail.

This patch however has been rejected upstream as Trond is apparently
working on a scheme that merges a lot of the DIO and buffered IO code in
NFS. That hasn't materialized yet and may not be suitable for RHEL6 in
any case. For now, I think this patch is a reasonable hack to make QEMU
work well on NFS with O_DIRECT.

IBM has tested this extensively with QEMU and reports a rather
significant performance boost with it. I also have a synthetic test that
shows a very significant perf boost with this patch.

----------------------[snip]------------------

Special readv/writev support for NFS O_DIRECT. Currently NFS
client O_DIRECT read/write iterates over individual elements
in the vector and submits the IO and waits for them to finish.
If individual IOsize is < r/wsize, each IO will be FILE_SYNC.
Server has to finish individual smaller IOs synchronous. This
causes serious performance degrade.

This patch is a special case to submit larger IOs from the client
only if all the IO buffers are page-aligned and each individual
IOs are page-multiple + total IO size is < r/wsize.

This is the common case for QEMU usage.

[Jeff: modified the patch according to David Howells' recommendations]

Signed-off-by: Badari Pulavarty <pbadari@us.ibm.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>