perf top browser: Adjust the browser indexes when refreshing
Patchwork-id: 36525
O-Subject: [PATCH RHEL6.2 BZ691343 087/142] perf top browser: Adjust the browser indexes when refreshing
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit 6435a5e39d3e01a1a73a925ed53ee18619b0a368
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed Feb 23 07:25:02 2011 -0300

description
-----------
This is not a problem when we're not at the bottom of the active symbols
list, so was not noticed, but at the end of the screen it falls apart.

Fix it by adjusting the ui_browser indexes according to the new number
of entries in the rb_tree and by seeking from the start of the rb_tree
to find the new symbol at the top of the screen.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>