perf tools: Improve support for sessions with multiple events
Patchwork-id: 36543
O-Subject: [PATCH RHEL6.2 BZ691343 103/142] perf tools: Improve support for sessions with multiple events
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit e248de331a452f8771eda6ed4bb30d92c82df28b
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Sat Mar 5 21:40:06 2011 -0300

description
-----------
By creating an perf_evlist out of the attributes in the perf.data file
header, so that we can use evlists and evsels when reading recorded
sessions in addition to when we record sessions.

More work is needed to allow tools to allow the user to select which
events are wanted when browsing sessions, be it just one or a subset of
them, aggregated or showed at the same time but with different
indications on the UI to allow seeing workloads thru different views at
the same time.

But the overall goal/trend is to more uniformly use evsels and evlists.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>