perf symbol: Move sym_entry->skip to symbol->ignore
Patchwork-id: 36555
O-Subject: [PATCH RHEL6.2 BZ691343 113/142] perf symbol: Move sym_entry->skip to symbol->ignore
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit 171b3be9c42e97cd4530706654242f6a3efb6ac3
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Fri Mar 11 13:36:01 2011 -0300

description
-----------
While going thru each of the sym_entry fields looking to reduce it to
the set of entries needed when in an active symbols list, 'skip' should
really be in symbol, as we set it when loading the symtab.

And the space used by the basic symbol allocation remains the same as
we had 5 bytes of padding.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>