perf tools: Version incorrect with some versions of grep
Patchwork-id: 36561
O-Subject: [PATCH RHEL6.2 BZ691343 122/142] perf tools: Version incorrect with some versions of grep
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit 58d406ed6a5f1ca4bc1dba5390b718c67847fa5f
Author: Josh Hunt <johunt@akamai.com>
Date:   Tue Mar 15 19:16:40 2011 -0700

description
-----------
Some versions of grep don't treat '\s' properly. When building perf on such
systems and using a kernel tarball the perf version is unable to be determined
from the main kernel Makefile and the user is left with a version of '..'.
Replacing the use of '\s' with '[[:space:]]', which should work in all grep
versions, gives a usable version number.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>