perf symbols: Look at .dynsym again if .symtab not found
Patchwork-id: 36565
O-Subject: [PATCH RHEL6.2 BZ691343 127/142] perf symbols: Look at .dynsym again if .symtab not found
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit 60e4b10c5a27182bc8ce7435050a17cb61c94d00
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Tue Mar 22 15:42:14 2011 -0300

description
-----------
The original intent of the code was to repeat the search with
want_symtab = 0. But as the code stands now, we never hit the "default"
case of the switch statement. Which means we never repeat the search.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>