perf tools: Fixup exit path when not able to open events
Patchwork-id: 36568
O-Subject: [PATCH RHEL6.2 BZ691343 131/142] perf tools: Fixup exit path when not able to open events
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit c286c419c784c238cd699be37fec7a9acc30d89f
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Mon Mar 28 09:50:11 2011 -0300

description
-----------
We have to deal with the TUI mode in perf top, so that we don't end up
with a garbled screen when, say, a non root user on a machine with a
paranoid setting (the default) tries to use 'perf top'.

Introduce a ui__warning_paranoid() routine shared by top and record that
tells the user the valid values for /proc/sys/kernel/perf_event_paranoid.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>