perf evsel: Fix use of inherit
Patchwork-id: 36580
O-Subject: [PATCH RHEL6.2 BZ691343 140/142] perf evsel: Fix use of inherit
Bugzilla: 691343
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 691343
https://bugzilla.redhat.com/show_bug.cgi?id=691343

upstream status
---------------
commit 5d2cd90922c778908bd0cd669e572a5b5eafd737
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Apr 14 11:20:14 2011 -0300

description
-----------
perf stat doesn't mmap and its perfectly fine for it to use task-bound
counters with inheritance.

So set the attr.inherit on the caller and leave the syscall itself to
validate it.

When the mmap fails perf_evlist__mmap will just emit a warning if this
is the failure reason.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>