NFS: Don't use vm_map_ram() in readdir
Patchwork-id: 36610
O-Subject: [RHEL6.2 PATCH 01/13] NFS: Don't use vm_map_ram() in readdir
Bugzilla: 701788
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

vm_map_ram() is not available on NOMMU platforms, and causes trouble
on incoherrent architectures such as ARM when we access the page data
through both the direct and the virtual mapping.

The alternative is to use the direct mapping to access page data
for the case when we are not crossing a page boundary, but to copy
the data into a linear scratch buffer when we are accessing data
that spans page boundaries.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Tested-by: Marc Kleine-Budde <mkl@pengutronix.de>
Cc: stable@kernel.org  [2.6.37]

Upstream commit: 6650239a4b01077e80d5a4468562756d77afaa59

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>