xenpv: mask MWAIT cpuid feature
Patchwork-id: 36628
O-Subject: [RHEL6.2 PATCH] xenpv: mask MWAIT cpuid feature
Bugzilla: 712131
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>

bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=712131

We've hit a couple issues where an EC2 host has older hypervisor code
running on newer hardware, which exposes those newer cpu features to the
guests. This is dangerous because newer guest kernels (such as RHEL6) may
attempt to use those features or assume certain machine behaviors that it
won't get (since it's really a xen guest). Most recently we hit bug 709856,
which led to the opening of this bug.

One such place is the intel_idle driver which attempts to use
MWAIT/MONITOR, which are invalid ops for a xenpv guest. We need to mask
out the MWAIT feature to avoid that.

This is probably looking like a trend to reviewers. I hope this will be
the last round, and I've done a full analysis (written up in this bugzilla)
to try and confirm it. Also, this time I think upstream may be interested.
The pvops code has always masked a few features for defensive measures,
and I believe this feature fits the criteria.

I haven't tested this myself yet, but I'm under a time crunch to push it
out.  I'll be testing it as soon as my brew build completes, which is
brewing here http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3381891

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>