fix accounting bug on cross partition merges
Patchwork-id: 36631
O-Subject: [RHEL6 PATCH v2] block: fix accounting bug on cross partition merges
Bugzilla: 669363
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This version fixes the loop in is_same_part().

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=669363

This version use a totally different (much simpler) solution than the
previous one.

When a bio is merged into a request which belongs to a different partition
by ELEVATOR_FRONT_MERGE, the part->in_flight value which is decreased when
request is completed, is not the same as the one that was increase when the
request was added and we end up having both in_flight value corrupted.

This is done upstream by caching the partition but the backport of that
patch turned out to raise complicated kabi issues.
Here, it is fixed it by correctly updating the in_flight values when a
cross partition front merge occurs.

Regards,
Jerome

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>