GFS2: remove potential race in flock code
Patchwork-id: 36659
O-Subject: [RHEL 6.2] GFS2: kernel BUG at fs/gfs2/glock.c:173! running brawl
	w/flocks (bz #612608)
Bugzilla: 612608
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This is a very small patch which has been upstream for some time. We do
not have a reproducer for the originally reported issue, so the only
verification which can be done is to check that there are no regressions
as a result of this.

The patch simply adds a wait for the result of the glock demote
operation. This will always be fast since the results of demotes are
generated locally in the dlm, so that no network communication is
required. As a result this should not cause any performance issues.

Upstream patch is here:
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=0a33443b38746f35fc4acc8a5af6c7099e03ea40

This fixes bz #612608

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>