GFS2: force a log flush when invalidating the rindex glock
Patchwork-id: 36660
O-Subject: [RHEL 6.2 PATCH] GFS2: force a log flush when invalidating the
	rindex glock
Bugzilla: 702263
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

Right now, there is nothing that forces the log to get flushed when a node
drops its rindex glock so that another node can grow the filesystem. If the
log doesn't get flushed, GFS2 can corrupt the sd_log_le_rg list in the
following way:

A node puts an rgd on the list in rg_lo_add(), and then the rindex glock is
dropped so the other node can grow the filesystem. When the node reacquires the
rindex glock, that rgd gets deleted in clear_rgrpdi() before ever being
removed from the list by gfs2_log_flush().

This code simply forces a log flush when the rindex glock is invalidated,
solving the problem.

This fix has been submitted to Steve's upstream git tree
This fixes bz #702263

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>