setting CONFIG_THERMAL_HWMON as built in
Patchwork-id: 36782
O-Subject: [RHEL6.2 PATCH 2/2 v2] setting CONFIG_THERMAL_HWMON to 'y'
Bugzilla: 619044
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>

Resolves RHBZ 619044

Ensure that CONFIG_THERMAL_HWMON gets set to 'y' in order to enable some
important code in drivers/thermal/thermal_sys.c to be compiled into the
thermal_sys driver. It allows libsensors to access ACPI thermal zones.

In the BZ's comment #15, I mention that CONFIG_THERMAL_HWMON's Kconfig
definition makes it dependent on "HWMON=y || HWMON=THERMAL". And in
RHEL6, CONFIG_HWMON='m' and CONFIG_THERMAL='y'. So CONFIG_THERMAL_HWMON
doesn't end up getting set. (Note that the dependency is legit.)

There are two approaches one could take:  1) set CONFIG_HWMON='y', or
2) set CONFIG_THERMAL='m'.

Approach #1 changes CONFIG_HWMON from 'm' to 'y'. And approach #2
changes CONFIG_THERMAL, CONFIG_ACPI_FAN, CONFIG_ACPI_PROCESSOR, and
CONFIG_X86_POWERNOW_K7 from 'y' to 'm'.

It was pointed out that approach #2 introduces a potential risk. The
problem is that once the operating system asserts ACPI control, the
firmware may give up thermal control. At this point it is (per spec)
potentially up to the operating system to manage the thermal state of
the system. If the ACPI thermal driver is built as a module, and if we
fail to mount the root filesystem, we'll never load the thermal driver
and the system may overheat. Having the code built in reduces this
possibility, in that as long as we haven't entirely paniced we'll still
run the thermal code.

It's basically about risk reduction - in theory we'll load the module
fine, and everything will be okay. If we don't then the hardware should
shut down before there's any real thermal risk. And even if we have it
built in, the kernel may wedge in such a way that we don't run the thermal
code. But having it built in is safer than having it as a module, thus
I've gone with approach #1.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>