net: don't set VLAN_TAG_PRESENT for VLAN 0 frames
Patchwork-id: 36799
O-Subject: [RHEL 6.2 PATCH] net: don't set VLAN_TAG_PRESENT for VLAN 0 frames
Bugzilla: 706208
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=706208

Description:
When RHEL 6.1 is running as a guest on an ESX server, when loading the
802.1q driver and configuring an interface as a vlan trunk, the VM hangs
for a few minutes then reboots.  The hang/reboot is observed when 8021p
frames are passed to the guest OS.
The problem seems to occur because VLAN 0 frames are flagged with
VLAN_TAG_PRESENT and are thus passed to the VLAN subsystem.

Upstream has rewritten larger parts of the code, but the issue is still
not completely solved to my understanding. I talked to Andy and we deem
it to be saver to just not add VLAN_TAG_PRESENT to frames tagged with
VLAN 0.

Upstream Status:
not upstream

Brew Build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3391232

Test Status:
Unable to reproduce in-house. Successfully tested by Cisco.

  Stefan

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>